Move convert.py and make it a little more user friendly
